refactor: 拆分 writer.py 为 7 个职责单一模块并复用重复代码
将 1889 行的 writer.py 拆分为 7 个模块。提取 _ordered_pair、_manifest_target_lang、_render_paragraph_html、_assemble_plain_text 消除重复。writer.py 缩减为 87 行兼容入口。327 测试通过，五格式逐字节一致。
